Document vertical centering breaking backcompat
Essentially, slides have variable height now. This is not normally visible, but with transitions that have a non-transparent background, like `cube` or `page`, it is:




http://lab.hakim.se/reveal-js/?transition=cube




Before, slides had a fixed height, which means that presentations designed with that assumption in mind will look very different now.




Another aspect that I bumped into is that I used the slide's height to dynamically size images to fill the remainder of the space from the heading to the footer, like this (stripped to the essence):




          bigImage.style.height = section.offsetHeight - bigImage.offsetTop + 'px';




Alternatively, is there an easy way to restore the previous functionality?